refactor: move helpers to gimie.sources.common
