+++ gimie/sources/common/license.py
